first commit of new org

+++ cpp/stopwatch.hpp
@@ -1,13 +1,15 @@
 #ifndef STOPWATCH_HPP
 #define STOPWATCH_HPP
 #include <ctime>
-#include <sys/time.h>
 
 struct stopwatch  {
     struct timespec st;
     double elapsed_time;
 
-    stopwatch():st(),elapsed_time(0.) {
+    stopwatch():
+        st(),
+        elapsed_time(0.) 
+    {
     }
     ~stopwatch() {
     }
@@ -17,7 +19,7 @@ struct stopwatch  {
     void stop() noexcept {
         struct timespec fn;
         clock_gettime(CLOCK_MONOTONIC,&fn);
-        elapsed_time += (fn.tv_sec-st.tv_sec) + 1.e-9*(fn.tv_nsec-st.tv_nsec);
+        elapsed_time += double(fn.tv_sec-st.tv_sec) + 1.e-9*(fn.tv_nsec-st.tv_nsec);
     }
     void clear() noexcept {
         elapsed_time = 0.;
@@ -26,5 +28,4 @@ struct stopwatch  {
         return elapsed_time;
     }
 };
-
 #endif
